FakeCursor Component calculates and renders cursor
This component will take a TodoItem and an input reference. It will then use the TodoItems context to retrieve a CursorPositions list.

It will get the CursorPosition object for the item it is currently rendering, calculate where the fake cursor should be placed based on the input and render it.