Change event to 'notify-new-todo-item' event
Instead of having the socket event create an item, we use a 'notify-update' event. The item will be created by rest endpoint. The client will then have to emit this event to notify other users that there is a new item in the list.

Since a new item needs an id from backend this will be a lot easier to manage this way.